change kernel to 5.8.12 lts version

+++ kernel/drivers/md/bcache/alloc.c
@@ -87,7 +87,7 @@ void bch_rescale_priorities(struct cache_set *c, int sectors)
 {
 	struct cache *ca;
 	struct bucket *b;
-	unsigned long next = c->nbuckets * c->sb.bucket_size / 1024;
+	unsigned int next = c->nbuckets * c->sb.bucket_size / 1024;
 	unsigned int i;
 	int r;
